Develop

+++ back-end/delete.php
@@ -1,19 +1,27 @@
 <?php 
 include 'connexion.php';
+include 'auth.php';
 header("Access-Control-Allow-Origin: *");
 header("Access-Control-Allow-Methods: POST, GET, OPTIONS");
-header("Access-Control-Allow-Headers: Content-Type");
+header("Access-Control-Allow-Headers: Content-Type , Authorization");
 header("Content-Type: application/json");
 if ($_SERVER['REQUEST_METHOD'] === 'OPTIONS') {
     http_response_code(200);
     exit();
 }
-$data = json_decode(file_get_contents("php://input"), true);
-$id = $data['id'];
-$stmt = $connexion->prepare('DELETE FROM stagaire where id_stagaire= ?');
-if ($stmt->execute([$id])) {
-    echo json_encode(['success' => true, 'message' => 'User deleted']);
-} else {
-    echo json_encode(['success' => false, 'message' => 'Failed to delete user']);
+$user = verifyToken($connexion);
+if($user['role'] == 'admin'){
+
+    $data = json_decode(file_get_contents("php://input"), true);
+    $id = $data['id'];
+    $stmt = $connexion->prepare('DELETE FROM stagaire where id_stagaire= ?');
+    if ($stmt->execute([$id])) {
+        echo json_encode(['success' => true, 'message' => 'User deleted']);
+    } else {
+        echo json_encode(['success' => false, 'message' => 'Failed to delete user']);
+    }
+}
+else {
+    echo json_encode(['success' => false, 'message' => 'Permission denied']);
 }
